Rename proxy package to network
The package contains the NetworkSettingsConfigurator which also configures SSL
(and not just proxy settings).
